Added Transarent Blocks
Transparent blocks: Every side of the block has the same lightlevel, has 1 Display Entity instead of 6. This impacts performance of a very large scale, but is not noticable if you only have a few blocks.

Change Placed Block:
Chose what block is placed as the "core" of a custom block. (glass is recommended for transparent blocks.